requirements.txt
tesla-ce-provider>0.0.3
python-magic==0.4.22
